requirements.txt
apscheduler
mysql-connector
requests
lxml